[rename] 조회 기능 retrieve로 이동
